Refactor saiserver with single test fake
