+++ dataplane/standalone/saiserver/ports.go
@@ -20,8 +20,6 @@ import (
 	"net"
 
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
 
 	"github.com/openconfig/lemming/dataplane/standalone/saiserver/attrmgr"
@@ -310,77 +308,3 @@ func (port *port) Reset() {
 	port.nextEth = 1
 }
 
-func newHostif(mgr *attrmgr.AttrMgr, dataplane portDataplaneAPI, s *grpc.Server) *hostif {
-	p := &hostif{
-		mgr:       mgr,
-		dataplane: dataplane,
-	}
-	saipb.RegisterHostifServer(s, p)
-	return p
-}
-
-type hostif struct {
-	saipb.UnimplementedHostifServer
-	mgr       *attrmgr.AttrMgr
-	dataplane portDataplaneAPI
-}
-
-// CreateHostif creates a hostif interface (usually a tap interface).
-func (hostif *hostif) CreateHostif(ctx context.Context, req *saipb.CreateHostifRequest) (*saipb.CreateHostifResponse, error) {
-	id := hostif.mgr.NextID()
-
-	switch req.GetType() {
-	case saipb.HostifType_HOSTIF_TYPE_GENETLINK: // TODO: support this type
-		return &saipb.CreateHostifResponse{Oid: id}, nil
-	case saipb.HostifType_HOSTIF_TYPE_NETDEV:
-		portReq := &dpb.CreatePortRequest{
-			Id:   fmt.Sprint(id),
-			Type: fwdpb.PortType_PORT_TYPE_KERNEL,
-			Src: &dpb.CreatePortRequest_KernelDev{
-				KernelDev: string(req.GetName()),
-			},
-			Location: dpb.PortLocation_PORT_LOCATION_INTERNAL,
-		}
-		attrReq := &saipb.GetPortAttributeRequest{Oid: req.GetObjId(), AttrType: []saipb.PortAttr{saipb.PortAttr_PORT_ATTR_OPER_STATUS}}
-		p := &saipb.GetPortAttributeResponse{}
-		if err := hostif.mgr.PopulateAttributes(attrReq, p); err != nil {
-			return nil, err
-		}
-		if p.GetAttr().GetOperStatus() != saipb.PortOperStatus_PORT_OPER_STATUS_NOT_PRESENT {
-			portReq.ExternalPort = fmt.Sprint(req.GetObjId())
-		}
-		if _, err := hostif.dataplane.CreatePort(ctx, portReq); err != nil {
-			if err != nil {
-				return nil, err
-			}
-		}
-		attr := &saipb.HostifAttribute{
-			OperStatus: proto.Bool(true),
-		}
-		hostif.mgr.StoreAttributes(id, attr)
-	default:
-		return nil, status.Errorf(codes.InvalidArgument, "unknown type %v", req.GetType())
-	}
-	return &saipb.CreateHostifResponse{Oid: id}, nil
-}
-
-// SetHostifAttribute sets the attributes in the request.
-func (hostif *hostif) SetHostifAttribute(ctx context.Context, req *saipb.SetHostifAttributeRequest) (*saipb.SetHostifAttributeResponse, error) {
-	if req.OperStatus != nil {
-		stateReq := &fwdpb.PortStateRequest{
-			ContextId: &fwdpb.ContextId{Id: hostif.dataplane.ID()},
-			PortId:    &fwdpb.PortId{ObjectId: &fwdpb.ObjectId{Id: fmt.Sprint(req.GetOid())}},
-		}
-		stateReq.Operation = &fwdpb.PortInfo{
-			AdminStatus: fwdpb.PortState_PORT_STATE_DISABLED_DOWN,
-		}
-		if req.GetOperStatus() {
-			stateReq.Operation.AdminStatus = fwdpb.PortState_PORT_STATE_ENABLED_UP
-		}
-		_, err := hostif.dataplane.PortState(ctx, stateReq)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return nil, nil
-}
